Moved facade to multivisualisation package.
