+++ abixen-platform-business-intelligence-service/src/main/java/com/abixen/platform/service/businessintelligence/configuration/PlatformModulesPackages.java
@@ -29,8 +29,6 @@ public class PlatformModulesPackages {
 
     public static final String CONTROLLER = MAIN + ".controller";
 
-    public static final String FACADE = MAIN + ".facade";
-
     public static final String CLIENT = COMMON + ".client";
 
     public static final String INTEGRATION = COMMON + ".integration";
